basic stats
I created seperate packages for items, enemies and the player stats. i made a superclass for each package, then its subclasses are the specific items. all logic i did was done to the game controller, which was basically displaying the data to the screen, which means we can mess arond with the data types created. if having too many packages makes it more complex, we can technically move it all under the app package; was just trying to make it look more cool lol.